+++ libseccomp/src/error.rs
@@ -10,9 +10,6 @@ use std::ops::Deref;
 
 pub(crate) type Result<T> = std::result::Result<T, SeccompError>;
 
-// ParseError message
-const PARSE_ERROR: &str = "Parse error by invalid argument";
-
 /// Errnos returned by the libseccomp API.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 #[non_exhaustive]
@@ -60,6 +57,24 @@ impl SeccompErrno {
             ESRCH => "Unable to load the filter due to thread issues",
         }
     }
+
+    fn to_sysrawrc(self) -> i32 {
+        use SeccompErrno::*;
+
+        match self {
+            EACCES => libc::EACCES,
+            ECANCELED => libc::ECANCELED,
+            EDOM => libc::EDOM,
+            EEXIST => libc::EEXIST,
+            EFAULT => libc::EFAULT,
+            EINVAL => libc::EINVAL,
+            ENOENT => libc::ENOENT,
+            ENOMEM => libc::ENOMEM,
+            EOPNOTSUPP => libc::EOPNOTSUPP,
+            ERANGE => libc::ERANGE,
+            ESRCH => libc::ESRCH,
+        }
+    }
 }
 
 impl fmt::Display for SeccompErrno {
@@ -74,8 +89,14 @@ impl fmt::Display for SeccompErrno {
 pub(crate) enum ErrorKind {
     /// An error that represents error code on failure of the libseccomp API.
     Errno(SeccompErrno),
-    /// A parse error occurred while trying to convert a value.
-    ParseError,
+    /// A system's raw error code.
+    SysRawRc(i32),
+    /// An invalid Architecture.
+    InvalidArch(u32),
+    /// An invalid Action.
+    InvalidAction(u32),
+    /// An invalid string in from_str.
+    FromStr(String),
     /// A lower-level error that is caused by an error from a lower-level module.
     Source,
     /// A custom error that does not fall under any other error kind.
@@ -138,12 +159,7 @@ impl SeccompError {
             libc::EOPNOTSUPP => SeccompErrno::EOPNOTSUPP,
             libc::ERANGE => SeccompErrno::ERANGE,
             libc::ESRCH => SeccompErrno::ESRCH,
-            _ => {
-                return Self::with_msg(format!(
-                    "libseccomp-rs error: errno {} not handled.",
-                    raw_errno,
-                ))
-            }
+            _ => return Self::new(ErrorKind::SysRawRc(raw_errno)),
         };
         Self::new(ErrorKind::Errno(seccomp_errno))
     }
@@ -157,12 +173,70 @@ impl SeccompError {
         }
     }
 
+    /// Query the system's raw error code returned when something goes wrong
+    /// in the libc and the kernel.
+    ///
+    /// This function will be useful for users who want to extract the system's
+    /// error code directly returned by [`ScmpFilterAttr::ApiSysRawRc`](`crate::ScmpFilterAttr::ApiSysRawRc`)
+    /// , or get the errno returned by the libseccomp API as a negative integer rather than [`SeccompErrno`].
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # use libseccomp::*;
+    /// # use std::os::fd::*;
+    /// let mut ctx = ScmpFilterContext::new(ScmpAction::Allow)?;
+    /// ctx.set_api_sysrawrc(true)?;
+    /// match ctx.export_pfc(unsafe { OwnedFd::from_raw_fd(-2) }) {
+    ///     Err(e) => {
+    ///         eprintln!("Error: {e}");
+    ///         if let Some(sys) = e.sysrawrc() {
+    ///             eprintln!("The system's raw error code: {sys}");
+    ///             assert_eq!(sys, -libc::EBADF);
+    ///         }
+    ///     }
+    ///     _ => println!("No error"),
+    /// }
+    /// # Ok::<(), Box<dyn std::error::Error>>(())
+    /// ```
+    pub fn sysrawrc(&self) -> Option<i32> {
+        match self.kind {
+            ErrorKind::SysRawRc(rc) => Some(rc),
+            ErrorKind::Errno(errno) => Some(-errno.to_sysrawrc()),
+            _ => None,
+        }
+    }
+
+    /// Returns the raw ffi value of an unsupported Action/Arch.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # use libseccomp::*;
+    /// let mut ctx = ScmpFilterContext::new(ScmpAction::Allow)?;
+    /// if let Err(err) = ctx.get_act_default() {
+    ///     println!("{:#?}", err.raw_ffi_value())
+    /// }
+    /// # Ok::<(), Box<dyn std::error::Error>>(())
+    /// ```
+    pub fn raw_ffi_value(&self) -> Option<u32> {
+        match self.kind {
+            ErrorKind::InvalidArch(v) | ErrorKind::InvalidAction(v) => Some(v),
+            _ => None,
+        }
+    }
+
     fn msg(&self) -> Cow<'_, str> {
         match &self.kind {
             ErrorKind::Errno(e) => e.strerror().into(),
-            ErrorKind::Common(s) => s.deref().into(),
-            ErrorKind::ParseError => PARSE_ERROR.into(),
+            ErrorKind::SysRawRc(e) => {
+                format!("The system's raw error code({}) was returned", e).into()
+            }
+            ErrorKind::InvalidArch(_) => "Parse error by invalid architecture".into(),
+            ErrorKind::InvalidAction(_) => "Parse error by invalid action".into(),
+            ErrorKind::FromStr(s) => format!("Error while parsing '{s}'").into(),
             ErrorKind::Source => self.source.as_ref().unwrap().to_string().into(),
+            ErrorKind::Common(s) => s.deref().into(),
         }
     }
 }
@@ -281,10 +355,6 @@ mod tests {
             SeccompError::from_errno(-libc::ESRCH).msg(),
             SeccompErrno::ESRCH.strerror()
         );
-        assert_eq!(
-            SeccompError::from_errno(-libc::EPIPE).msg(),
-            format!("libseccomp-rs error: errno {} not handled.", -libc::EPIPE)
-        );
 
         // Common
         assert_eq!(
@@ -292,13 +362,34 @@ mod tests {
             TEST_ERR_MSG
         );
 
-        // ParseError
-        assert_eq!(SeccompError::new(ParseError).msg(), PARSE_ERROR);
-
         // Source
         assert_eq!(
             SeccompError::with_source(Source, null_err).msg(),
             NULL_ERR_MSG
+        );
+
+        // SysRawRc
+        assert_eq!(
+            SeccompError::from_errno(-libc::EPIPE).msg(),
+            format!("The system's raw error code({}) was returned", -libc::EPIPE)
+        );
+
+        // InvalidArch
+        assert_eq!(
+            SeccompError::new(InvalidArch(123)).msg(),
+            "Parse error by invalid architecture",
+        );
+
+        // InvalidAction
+        assert_eq!(
+            SeccompError::new(InvalidAction(123)).msg(),
+            "Parse error by invalid action",
+        );
+
+        // FromStr
+        assert_eq!(
+            SeccompError::new(FromStr("SCMP".to_string())).msg(),
+            "Error while parsing 'SCMP'",
         );
     }
 
@@ -317,6 +408,76 @@ mod tests {
     }
 
     #[test]
+    fn test_with_msg() {
+        assert_eq!(SeccompError::with_msg(TEST_ERR_MSG).msg(), TEST_ERR_MSG);
+        assert!(SeccompError::with_msg(TEST_ERR_MSG).source().is_none());
+    }
+
+    #[test]
+    fn test_with_msg_and_source() {
+        let null_err = CString::new(TEST_NULL_STR).unwrap_err();
+
+        assert_eq!(
+            SeccompError::with_msg_and_source(TEST_ERR_MSG, null_err.clone()).msg(),
+            TEST_ERR_MSG
+        );
+        assert!(SeccompError::with_msg_and_source(TEST_ERR_MSG, null_err)
+            .source()
+            .is_some());
+    }
+
+    #[test]
+    fn test_errno() {
+        assert_eq!(
+            SeccompError::from_errno(-libc::EACCES).errno().unwrap(),
+            SeccompErrno::EACCES
+        );
+        assert!(SeccompError::from_errno(libc::EBADFD).errno().is_none());
+    }
+
+    #[test]
+    fn test_sysrawrc() {
+        let tests = &[
+            // The EBADFD is not handled by SeccompErrno
+            libc::EBADFD,
+            // The following errnos are handled by SeccompErrno
+            libc::EACCES,
+            libc::ECANCELED,
+            libc::EDOM,
+            libc::EEXIST,
+            libc::EFAULT,
+            libc::EINVAL,
+            libc::ENOENT,
+            libc::ENOMEM,
+            libc::EOPNOTSUPP,
+            libc::ERANGE,
+            libc::ESRCH,
+        ];
+
+        for errno in tests {
+            assert_eq!(SeccompError::from_errno(-errno).sysrawrc().unwrap(), -errno);
+        }
+        assert!(SeccompError::with_msg("no errno").sysrawrc().is_none());
+    }
+
+    #[test]
+    fn test_raw_ffi_value() {
+        assert_eq!(
+            SeccompError::new(InvalidArch(123)).raw_ffi_value().unwrap(),
+            123
+        );
+        assert_eq!(
+            SeccompError::new(InvalidAction(123))
+                .raw_ffi_value()
+                .unwrap(),
+            123
+        );
+        assert!(SeccompError::new(Common("".into()))
+            .raw_ffi_value()
+            .is_none());
+    }
+
+    #[test]
     fn test_from() {
         let null_err = CString::new(TEST_NULL_STR).unwrap_err();
         let scmp_err = SeccompError::from(null_err.clone());
@@ -328,6 +489,12 @@ mod tests {
     #[test]
     fn test_display() {
         let null_err = CString::new(TEST_NULL_STR).unwrap_err();
+
+        // fmt::Display for SeccompErrno
+        assert_eq!(
+            format!("{}", SeccompErrno::EACCES),
+            SeccompErrno::EACCES.strerror()
+        );
 
         // Errno without source
         assert_eq!(
@@ -361,17 +528,6 @@ mod tests {
             format!("{} caused by: {}", TEST_ERR_MSG, NULL_ERR_MSG)
         );
 
-        // Parse without source
-        assert_eq!(format!("{}", SeccompError::new(ParseError)), PARSE_ERROR);
-        // Parse with source
-        assert_eq!(
-            format!(
-                "{}",
-                SeccompError::with_source(ParseError, null_err.clone())
-            ),
-            format!("{} caused by: {}", PARSE_ERROR, NULL_ERR_MSG)
-        );
-
         // Source
         assert_eq!(
             format!("{}", SeccompError::with_source(ErrorKind::Source, null_err)),
@@ -380,7 +536,6 @@ mod tests {
     }
 
     #[test]
-    #[cfg(not(any(msrv_compat_1_52, msrv_compat_1_53)))]
     fn test_debug() {
         let null_err = CString::new(TEST_NULL_STR).unwrap_err();